requirements.txt
# Kivy for the mobile app UI
kivy==2.1.0
kivymd==1.1.1

# OpenCV for eye tracking and image processing
opencv-python==4.7.0.72

# SpeechRecognition for voice-to-text conversion
SpeechRecognition==3.9.0

# gTTS for text-to-speech conversion
gtts==2.3.2

# Pyrebase for Firebase integration
pyrebase4==4.5.0

# OpenAI API for LLM integration (optional, if needed)
openai==0.27.0

# Additional utilities
numpy==1.24.2